create 2 dto (response & request), solve lazy initiliaation problem, (Actor & Address & Store) endpoint

+++ src/main/java/com/iti/sakilaapi/repository/implementation/ActorRepositoryImpl.java
@@ -4,7 +4,7 @@ import com.iti.sakilaapi.model.entity.Actor;
 import com.iti.sakilaapi.repository.TransactionalEntityManager;
 import com.iti.sakilaapi.repository.interfaces.ActorRepository;
 
-public class ActorRepositoryImpl extends BaseEntityRepositoryImpl<Actor, Short> implements ActorRepository {
+public class ActorRepositoryImpl extends BaseEntityRepositoryImpl<Actor, Integer> implements ActorRepository {
     public ActorRepositoryImpl() {
         super(new TransactionalEntityManager(), Actor.class);
     }
