+++ src/main/java/com/iti/sakilaapi/repository/implementation/CategoryRepositoryImpl.java
@@ -4,7 +4,7 @@ import com.iti.sakilaapi.model.entity.Category;
 import com.iti.sakilaapi.repository.TransactionalEntityManager;
 import com.iti.sakilaapi.repository.interfaces.CategoryRepository;
 
-public class CategoryRepositoryImpl extends BaseEntityRepositoryImpl<Category, Short> implements CategoryRepository {
+public class CategoryRepositoryImpl extends BaseEntityRepositoryImpl<Category, Integer> implements CategoryRepository {
     public CategoryRepositoryImpl() {
         super(new TransactionalEntityManager(), Category.class);
     }
