+++ src/main/java/com/iti/sakilaapi/repository/implementation/CountryRepositoryImpl.java
@@ -4,7 +4,7 @@ import com.iti.sakilaapi.model.entity.Country;
 import com.iti.sakilaapi.repository.TransactionalEntityManager;
 import com.iti.sakilaapi.repository.interfaces.CountryRepository;
 
-public class CountryRepositoryImpl extends BaseEntityRepositoryImpl<Country, Short> implements CountryRepository {
+public class CountryRepositoryImpl extends BaseEntityRepositoryImpl<Country, Integer> implements CountryRepository {
     public CountryRepositoryImpl() {
         super(new TransactionalEntityManager(), Country.class);
     }
