+++ src/main/java/com/iti/sakilaapi/repository/implementation/FilmTextRepositoryImpl.java
@@ -4,7 +4,7 @@ import com.iti.sakilaapi.model.entity.FilmText;
 import com.iti.sakilaapi.repository.TransactionalEntityManager;
 import com.iti.sakilaapi.repository.interfaces.FilmTextRepository;
 
-public class FilmTextRepositoryImpl extends BaseEntityRepositoryImpl<FilmText, Short> implements FilmTextRepository {
+public class FilmTextRepositoryImpl extends BaseEntityRepositoryImpl<FilmText, Integer> implements FilmTextRepository {
     public FilmTextRepositoryImpl() {
         super(new TransactionalEntityManager(), FilmText.class);
     }
