+++ src/main/java/com/iti/sakilaapi/repository/implementation/LanguageRepositoryImpl.java
@@ -4,7 +4,7 @@ import com.iti.sakilaapi.model.entity.Language;
 import com.iti.sakilaapi.repository.TransactionalEntityManager;
 import com.iti.sakilaapi.repository.interfaces.LanguageRepository;
 
-public class LanguageRepositoryImpl extends BaseEntityRepositoryImpl<Language, Short> implements LanguageRepository {
+public class LanguageRepositoryImpl extends BaseEntityRepositoryImpl<Language, Integer> implements LanguageRepository {
     public LanguageRepositoryImpl() {
         super(new TransactionalEntityManager(), Language.class);
     }
