+++ src/main/java/com/iti/sakilaapi/repository/implementation/PaymentRepositoryImpl.java
@@ -4,7 +4,7 @@ import com.iti.sakilaapi.model.entity.Payment;
 import com.iti.sakilaapi.repository.TransactionalEntityManager;
 import com.iti.sakilaapi.repository.interfaces.PaymentRepository;
 
-public class PaymentRepositoryImpl extends BaseEntityRepositoryImpl<Payment, Short> implements PaymentRepository {
+public class PaymentRepositoryImpl extends BaseEntityRepositoryImpl<Payment, Integer> implements PaymentRepository {
     public PaymentRepositoryImpl() {
         super(new TransactionalEntityManager(), Payment.class);
     }
