+++ src/main/java/com/iti/sakilaapi/repository/implementation/StaffRepositoryImpl.java
@@ -4,7 +4,7 @@ import com.iti.sakilaapi.model.entity.Staff;
 import com.iti.sakilaapi.repository.TransactionalEntityManager;
 import com.iti.sakilaapi.repository.interfaces.StaffRepository;
 
-public class StaffRepositoryImpl extends BaseEntityRepositoryImpl<Staff, Short> implements StaffRepository {
+public class StaffRepositoryImpl extends BaseEntityRepositoryImpl<Staff, Integer> implements StaffRepository {
     public StaffRepositoryImpl() {
         super(new TransactionalEntityManager(), Staff.class);
     }
