+++ src/main/java/com/iti/sakilaapi/repository/implementation/StoreRepositoryImpl.java
@@ -4,7 +4,7 @@ import com.iti.sakilaapi.model.entity.Store;
 import com.iti.sakilaapi.repository.TransactionalEntityManager;
 import com.iti.sakilaapi.repository.interfaces.StoreRepository;
 
-public class StoreRepositoryImpl extends BaseEntityRepositoryImpl<Store, Short> implements StoreRepository {
+public class StoreRepositoryImpl extends BaseEntityRepositoryImpl<Store, Integer> implements StoreRepository {
     public StoreRepositoryImpl() {
         super(new TransactionalEntityManager(), Store.class);
     }
