+++ src/main/java/com/iti/sakilaapi/repository/interfaces/ActorRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Actor;
 
-public interface ActorRepository extends BaseEntityRepository<Actor, Short> {
+public interface ActorRepository extends BaseEntityRepository<Actor, Integer> {
 }
