+++ src/main/java/com/iti/sakilaapi/repository/interfaces/AddressRepository.java
@@ -2,5 +2,6 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Address;
 
-public interface AddressRepository extends BaseEntityRepository<Address, Short> {
+public interface AddressRepository extends BaseEntityRepository<Address, Integer> {
+//    Address saveAddress(Address address);
 }
