+++ src/main/java/com/iti/sakilaapi/repository/interfaces/CategoryRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Category;
 
-public interface CategoryRepository extends BaseEntityRepository<Category, Short> {
+public interface CategoryRepository extends BaseEntityRepository<Category, Integer> {
 }
