+++ src/main/java/com/iti/sakilaapi/repository/interfaces/CityRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.City;
 
-public interface CityRepository extends BaseEntityRepository<City, Short> {
+public interface CityRepository extends BaseEntityRepository<City, Integer> {
 }
