+++ src/main/java/com/iti/sakilaapi/repository/interfaces/CountryRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Country;
 
-public interface CountryRepository extends BaseEntityRepository<Country, Short> {
+public interface CountryRepository extends BaseEntityRepository<Country, Integer> {
 }
