+++ src/main/java/com/iti/sakilaapi/repository/interfaces/CustomerRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Customer;
 
-public interface CustomerRepository extends BaseEntityRepository<Customer, Short> {
+public interface CustomerRepository extends BaseEntityRepository<Customer, Integer> {
 }
