+++ src/main/java/com/iti/sakilaapi/repository/interfaces/FilmActorRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.FilmActor;
 
-public interface FilmActorRepository extends BaseEntityRepository<FilmActor, Short> {
+public interface FilmActorRepository extends BaseEntityRepository<FilmActor, Integer> {
 }
