+++ src/main/java/com/iti/sakilaapi/repository/interfaces/FilmCategoryRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.FilmCategory;
 
-public interface FilmCategoryRepository extends BaseEntityRepository<FilmCategory, Short> {
+public interface FilmCategoryRepository extends BaseEntityRepository<FilmCategory, Integer> {
 }
