+++ src/main/java/com/iti/sakilaapi/repository/interfaces/FilmRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Film;
 
-public interface FilmRepository extends BaseEntityRepository<Film, Short> {
+public interface FilmRepository extends BaseEntityRepository<Film, Integer> {
 }
