+++ src/main/java/com/iti/sakilaapi/repository/interfaces/FilmTextRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.FilmText;
 
-public interface FilmTextRepository extends BaseEntityRepository<FilmText, Short> {
+public interface FilmTextRepository extends BaseEntityRepository<FilmText, Integer> {
 }
