+++ src/main/java/com/iti/sakilaapi/repository/interfaces/InventoryRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Inventory;
 
-public interface InventoryRepository extends BaseEntityRepository<Inventory, Short> {
+public interface InventoryRepository extends BaseEntityRepository<Inventory, Integer> {
 }
