+++ src/main/java/com/iti/sakilaapi/repository/interfaces/LanguageRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Language;
 
-public interface LanguageRepository extends BaseEntityRepository<Language, Short> {
+public interface LanguageRepository extends BaseEntityRepository<Language, Integer> {
 }
