+++ src/main/java/com/iti/sakilaapi/repository/interfaces/PaymentRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Payment;
 
-public interface PaymentRepository extends BaseEntityRepository<Payment, Short> {
+public interface PaymentRepository extends BaseEntityRepository<Payment, Integer> {
 }
