+++ src/main/java/com/iti/sakilaapi/repository/interfaces/RentalRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Rental;
 
-public interface RentalRepository extends BaseEntityRepository<Rental, Short> {
+public interface RentalRepository extends BaseEntityRepository<Rental, Integer> {
 }
