+++ src/main/java/com/iti/sakilaapi/repository/interfaces/StaffRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Staff;
 
-public interface StaffRepository extends BaseEntityRepository<Staff, Short> {
+public interface StaffRepository extends BaseEntityRepository<Staff, Integer> {
 }
