+++ src/main/java/com/iti/sakilaapi/repository/interfaces/StoreRepository.java
@@ -2,5 +2,5 @@ package com.iti.sakilaapi.repository.interfaces;
 
 import com.iti.sakilaapi.model.entity.Store;
 
-public interface StoreRepository extends BaseEntityRepository<Store, Short> {
+public interface StoreRepository extends BaseEntityRepository<Store, Integer> {
 }
